refactor: migrate hooks (use-issue-properties) from web/app/ce to web/app/core
